Refactor timesys module.
