+++ va/timesys/__init__.py
@@ -1,1 +1,3 @@
-"""Timesys subpackage."""
+"""Import Timing Simulation Class."""
+
+from .time_simul import TimingSimulation
