--- va/timesys/time_simul/__init__.py
@@ -1,3 +1,0 @@
-"""Import Timing Simulation Class."""
-
-from .time_simul import TimingSimulation
